backend/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
pandas==1.5.0
matplotlib==3.6.0
numpy==1.23.4
scikit-learn==1.1.2
Pillow==9.3.0
seaborn==0.12.1